+++ CHANGELOG.md
@@ -11,7 +11,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 :microscope: - experimental
 
 ## [0.52.0]
-- :rocket: added _I wait for network idle for {int} ms_ step
+- :rocket: added _I wait for network idle {playwrightTimeout}_ step
 
 ## [0.51.0]
 Breaking change: